Implementation moved to a new file imagebufalgo_poisson.cpp. Testsuite added for smooth completion.
