+++ src/CMakeLists.txt
@@ -260,7 +260,7 @@ endif()
 
 # List all the individual testsuite tests here:
 oiio_add_tests (ico gpsread misnamed-file
-                oiiotool oiiotool-fixnan 
+                oiiotool oiiotool-poisson oiiotool-fixnan 
                 sgi rla psd dpx png
                 texture-fill texture-gray texture-grid
                 texture-missing texture-overscan
